+++ astropy/stats/funcs.py
@@ -18,11 +18,24 @@ from ..extern.six.moves import xrange
 
 __all__ = ['sigma_clip', 'binom_conf_interval', 'binned_binom_proportion',
            'median_absolute_deviation', 'biweight_location',
-           'biweight_midvariance', 'signal_to_noise_oir_ccd', 'bootstrap']
-
+           'biweight_midvariance', 'signal_to_noise_oir_ccd', 'bootstrap',
+           'mad_std', 'gaussian_fwhm_to_sigma', 'gaussian_sigma_to_fwhm']
 
 __doctest_skip__ = ['binned_binom_proportion']
 __doctest_requires__ = {'binom_conf_interval': ['scipy.special']}
+
+
+gaussian_sigma_to_fwhm = 2.0 * np.sqrt(2.0 * np.log(2.0))
+"""
+Factor with which to multiply Gaussian 1-sigma standard deviation(s) to
+convert them to full width at half maximum(s).
+"""
+
+gaussian_fwhm_to_sigma = 1. / gaussian_sigma_to_fwhm
+"""
+Factor with which to multiply Gaussian full width at half maximum(s) to
+convert them to 1-sigma standard deviation(s).
+"""
 
 
 def sigma_clip(data, sig=3, iters=1, cenfunc=np.ma.median, varfunc=np.var,
@@ -871,3 +884,43 @@ def bootstrap(data, bootnum=100, samples=None, bootfunc=None):
             boot[i] = bootfunc(data[bootarr])
 
     return boot
+
+
+def mad_std(data):
+    """
+    Calculate a robust standard deviation using the `median absolute
+    deviation (MAD)
+    <http://en.wikipedia.org/wiki/Median_absolute_deviation>`_.
+
+    The standard deviation estimator is given by:
+
+    .. math::
+
+        \\sigma \\approx \\frac{\\textrm{MAD}}{\Phi^{-1}(3/4)} \\approx 1.4826 \ \\textrm{MAD}
+
+    where :math:`\Phi^{-1}(P)` is the normal inverse cumulative
+    distribution function evaulated at probability :math:`P = 3/4`.
+
+    Parameters
+    ----------
+    data : array-like
+        Data array or object that can be converted to an array.
+
+    Returns
+    -------
+    result : float
+        The robust standard deviation of the data.
+
+    Examples
+    --------
+    >>> from astropy.stats import mad_std
+    >>> from astropy.utils import NumpyRNGContext
+    >>> from numpy.random import normal
+    >>> with NumpyRNGContext(12345):
+    ...     data = normal(5, 2, size=(100, 100))
+    ...     mad_std(data)    # doctest: +FLOAT_CMP
+    2.02327646594
+    """
+
+    # NOTE: 1. / scipy.stats.norm.ppf(0.75) = 1.482602218505602
+    return median_absolute_deviation(data) * 1.482602218505602